ai-voice-agent-back/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv